moved TapGameUtils up to GameUtils
