+++ GameUtils/nettalk.data
