--- TapGameController/helpers.py
@@ -1,11 +1,0 @@
-"""
-This is a basic helper module for the GameController class
-"""
-
-def get_answer():
-    """Get an answer."""
-    return True
-
-def get_false():
-    """Always returns false"""
-    return not True
