+++ scripts/start_tap_game_controller.py
